Add files via upload
Ignore this file. This is the updated version of the original PyTorch-CNN-Model.ipynb. 
The update was made to save the entire model rather than just the weights. The change that was made:
torch.save(model, "ibm_ai_capstone_project_full.pth") -> torch.save(model.state_dict(), model_name)
This was needed for comparative analysis.